Port src/marks/area from JS to TS (stage 1)
Inline the hand-written area.d.ts types into a single area.ts source.
Behaviour is unchanged; runtime helpers imported from companion .js
modules carry @ts-ignore until those modules are themselves ported.
